Modify MemoryTableFactory
