+++ CMakeLists.txt
@@ -63,6 +63,7 @@ add_subdirectory(libmptstate)
 add_subdirectory(libexecutivecontext)
 add_subdirectory(libblockverifier)
 add_subdirectory(libstorage)
+add_subdirectory(libstoragestate)
 add_subdirectory(libtxpool)
 add_subdirectory(libconsensus)
 add_subdirectory(libblockchain)
