+++ libblockverifier/ExecutiveContext.cpp
@@ -103,7 +103,7 @@ std::shared_ptr<storage::Table> ExecutiveContext::getTable(const Address& addres
     std::string tableName = "_contract_data_" + address.hex() + "_";
     TableFactoryPrecompiled::Ptr tableFactoryPrecompiled =
         std::dynamic_pointer_cast<TableFactoryPrecompiled>(getPrecompiled(Address(0x1001)));
-    return tableFactoryPrecompiled->getmemoryTableFactory()->openTable(h256(), 0, tableName);
+    return tableFactoryPrecompiled->getmemoryTableFactory()->openTable(tableName);
 }
 
 std::shared_ptr<dev::eth::StateFace> ExecutiveContext::getState()
