+++ libstorage/MemoryTableFactory.h
@@ -38,9 +38,9 @@ public:
     MemoryTableFactory();
     virtual ~MemoryTableFactory() {}
 
-    Table::Ptr openTable(h256 blockHash, int64_t num, const std::string& table) override;
-    Table::Ptr createTable(h256 blockHash, int64_t num, const std::string& tableName,
-        const std::string& keyField, const std::string& valueField) override;
+    Table::Ptr openTable(const std::string& table) override;
+    Table::Ptr createTable(const std::string& tableName, const std::string& keyField,
+        const std::string& valueField) override;
 
     virtual Storage::Ptr stateStorage() { return m_stateStorage; }
     virtual void setStateStorage(Storage::Ptr stateStorage) { m_stateStorage = stateStorage; }
