+++ libstorage/Table.h
@@ -176,9 +176,9 @@ public:
 
     virtual ~StateDBFactory() {}
 
-    virtual Table::Ptr openTable(h256 blockHash, int64_t num, const std::string& table) = 0;
-    virtual Table::Ptr createTable(h256 blockHash, int64_t num, const std::string& tableName,
-        const std::string& keyField, const std::string& valueField) = 0;
+    virtual Table::Ptr openTable(const std::string& table) = 0;
+    virtual Table::Ptr createTable(const std::string& tableName, const std::string& keyField,
+        const std::string& valueField) = 0;
 };
 
 }  // namespace storage
