+++ test/CMakeLists.txt
@@ -85,6 +85,7 @@ target_link_libraries(test-fisco-bcos evmc)
 target_link_libraries(test-fisco-bcos interpreter)
 target_link_libraries(test-fisco-bcos p2p)
 target_link_libraries(test-fisco-bcos mptstate)
+target_link_libraries(test-fisco-bcos storagestate)
 target_link_libraries(test-fisco-bcos executivecontext)
 target_link_libraries(test-fisco-bcos txpool)
 target_link_libraries(test-fisco-bcos blockchain)
