+++ test/unittests/libstorage/test_MemoryTableFactory.cpp
@@ -58,9 +58,9 @@ BOOST_AUTO_TEST_CASE(open_Table)
     std::string tableName("t_test");
     std::string keyField("hash");
     std::string valueField("hash");
-    memoryDBFactory->createTable(blockHash, num, tableName, keyField, valueField);
+    memoryDBFactory->createTable(tableName, keyField, valueField);
     MemoryTable::Ptr db = std::dynamic_pointer_cast<MemoryTable>(
-        memoryDBFactory->openTable(h256(0x12345), 1, "t_test"));
+        memoryDBFactory->openTable("t_test"));
 }
 
 BOOST_AUTO_TEST_CASE(setBlockHash)
